Fine tuned service and resource classes

+++ src/main/java/com/asu/edu/cse564/resource/AssignmentResource.java
@@ -45,7 +45,7 @@ public class AssignmentResource {
         String jsonString = null;
         //if(!(gradeBooksUIInp.getGrade()==null))
         // We can check for presence of grade value as non-zero,but 0 can also be an update item so no check as of now
-        if(gid != 0 && sid != 0 || aid != 0) {
+        if((gid != 0 && sid != 0 && aid != 0) && (gradeBooksUIInp.getName() != null) ) {
             Assignment assignment = assignmentService.addAssignmentForStudentWithIdForGradeBook(gid, sid, aid, gradeBooksUIInp.getName() );
             if(assignment != null ) {
                 try {
@@ -56,14 +56,17 @@ public class AssignmentResource {
                 } 
                 locationURI = URI.create(context.getAbsolutePath().toString());
                 System.out.println(locationURI);
-                response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
+                response = Response.status(Response.Status.CREATED).location(locationURI).entity(jsonString).build();
             } else {
-                locationURI = URI.create(context.getAbsolutePath().toString());
-                response = Response.status(Response.Status.NO_CONTENT).location(locationURI).build();  //NO_CONTENT aa??
+                //locationURI = URI.create(context.getAbsolutePath().toString());
+                //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                response = Response.status(Response.Status.NOT_FOUND).build();  //NO_CONTENT aa??
             }
         } else {
-            locationURI = URI.create(context.getAbsolutePath().toString());
-            response = Response.status(Response.Status.NO_CONTENT).location(locationURI).build();
+            //locationURI = URI.create(context.getAbsolutePath().toString());
+            //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            String error = "Either an Id value is zero or Assignment Name is empty - Cant Create";
+            response = Response.status(Response.Status.BAD_REQUEST).entity(error).build();
         }
         return response;
     }
@@ -76,20 +79,24 @@ public class AssignmentResource {
     public Response deleteAssignmentForStudentWithIdForGradeBook(@PathParam("gid") int gid, @PathParam("sid") int sid, @PathParam("aid") int aid,GradeBooksUIInp gradeBooksUIInp) {
         Response response = null;
         URI locationURI;
-        if(gid != 0 && sid != 0 || aid != 0) {
+        if(gid != 0 && sid != 0 && aid != 0) {
             int status = assignmentService.deleteAssignmentForStudentWithIdForGradeBook(gid, sid, aid);
             if(status == 1) {
                 locationURI = URI.create(context.getAbsolutePath().toString());
                 System.out.println(locationURI);
-                response = Response.status(Response.Status.OK).location(locationURI).build();
+                //response = Response.status(Response.Status.NO_CONTENT).location(locationURI).build(); // OK + NOTHING TO RETURN
+                response = Response.status(Response.Status.NO_CONTENT).build(); // OK + NOTHING TO RETURN
             } else {
-                locationURI = URI.create(context.getAbsolutePath().toString());
-                response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                //locationURI = URI.create(context.getAbsolutePath().toString());
+                //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                response = Response.status(Response.Status.NOT_FOUND).build();  //NO_CONTENT aa??
             }
         }
         else {
-            locationURI = URI.create(context.getAbsolutePath().toString());
-            response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();            
+            //locationURI = URI.create(context.getAbsolutePath().toString());
+            //response = Response.status(Response.Status.BAD_REQUEST).location(locationURI).build();   //
+            String error = "Neither of the Id value can be zero - Cant Delete";
+            response = Response.status(Response.Status.BAD_REQUEST).entity(error).build();
         }
             
         return response;
@@ -119,12 +126,15 @@ public class AssignmentResource {
                 System.out.println(locationURI);
                 response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
             } else {
-                locationURI = URI.create(context.getAbsolutePath().toString());
-                response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                //locationURI = URI.create(context.getAbsolutePath().toString());
+                //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                response = Response.status(Response.Status.NOT_FOUND).build();  //NO_CONTENT aa??
             }
         } else {
-            locationURI = URI.create(context.getAbsolutePath().toString());
-            response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            //locationURI = URI.create(context.getAbsolutePath().toString());
+            //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            String error = "Neither of the Id value can be zero - Cant Update Grade";
+            response = Response.status(Response.Status.BAD_REQUEST).entity(error).build();
         }
         return response;
     }
@@ -132,14 +142,14 @@ public class AssignmentResource {
     
     
     @GET
-    @Consumes(MediaType.APPLICATION_JSON)
+    //@Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @Path("{aid}")
-    public Response getAssignmentForStudentWithIdForGradeBook(@PathParam("gid") int gid, @PathParam("sid") int sid, @PathParam("aid") int aid,GradeBooksUIInp gradeBooksUIInp) {
+    public Response getAssignmentForStudentWithIdForGradeBook(@PathParam("gid") int gid, @PathParam("sid") int sid, @PathParam("aid") int aid) {
         Response response = null;
         URI locationURI;
         String jsonString = null;
-        if(gid != 0 && sid != 0 || aid != 0) {
+        if(gid != 0 && sid != 0 && aid != 0) {
             // We can check for presence of grade item value as non-zero,but 0 can also be an update item so no check as of now 
             Assignment assignment = assignmentService.getAssignmentForStudentWithIdForGradeBook(gid, sid, aid);
             if(assignment != null ) {
@@ -151,14 +161,18 @@ public class AssignmentResource {
                 } 
                 locationURI = URI.create(context.getAbsolutePath().toString());
                 System.out.println(locationURI);
-                response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
+                //response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
+                response = Response.status(Response.Status.OK).entity(jsonString).build();
             } else {
-                locationURI = URI.create(context.getAbsolutePath().toString());
-                response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                //locationURI = URI.create(context.getAbsolutePath().toString());
+                //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                response = Response.status(Response.Status.NOT_FOUND).build();  //NO_CONTENT aa??
             }
         } else {
-            locationURI = URI.create(context.getAbsolutePath().toString());
-            response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            //locationURI = URI.create(context.getAbsolutePath().toString());
+            //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            String error = "Neither of the Id value can be zero - Cant Read";
+            response = Response.status(Response.Status.BAD_REQUEST).entity(error).build();
         }
         
         return response;
@@ -166,10 +180,10 @@ public class AssignmentResource {
     
     
     @GET
-    @Consumes(MediaType.APPLICATION_JSON)
+    //@Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     //@Path("{aid}")
-    public Response getAllAssignmentForStudentWithIdForGradeBook(@PathParam("gid") int gid, @PathParam("sid") int sid,GradeBooksUIInp gradeBooksUIInp) {
+    public Response getAllAssignmentForStudentWithIdForGradeBook(@PathParam("gid") int gid, @PathParam("sid") int sid) {
         Response response = null;
         URI locationURI;
         String jsonString = null;
@@ -185,14 +199,18 @@ public class AssignmentResource {
                 } 
                 locationURI = URI.create(context.getAbsolutePath().toString());
                 System.out.println(locationURI);
-                response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
+                //response = Response.status(Response.Status.OK).location(locationURI).entity(jsonString).build();
+                response = Response.status(Response.Status.OK).entity(jsonString).build();
             } else {
-                locationURI = URI.create(context.getAbsolutePath().toString());
-                response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                //locationURI = URI.create(context.getAbsolutePath().toString());
+                //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();  //NO_CONTENT aa??
+                response = Response.status(Response.Status.NOT_FOUND).build();  //NO_CONTENT aa??
             }
         } else {
-            locationURI = URI.create(context.getAbsolutePath().toString());
-            response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            //locationURI = URI.create(context.getAbsolutePath().toString());
+            //response = Response.status(Response.Status.NOT_FOUND).location(locationURI).build();
+            String error = "Neither of the Id value can be zero - Cant Read";
+            response = Response.status(Response.Status.BAD_REQUEST).entity(error).build();
         }
         
         return response;
